restructuration of structures and correction syntax about them

+++ philo/headers/init.h
@@ -6,7 +6,7 @@
 /*   By: fberthou <fberthou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/02 18:17:59 by fberthou          #+#    #+#             */
-/*   Updated: 2024/05/13 15:58:41 by fberthou         ###   ########.fr       */
+/*   Updated: 2024/07/12 10:29:09 by fberthou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@
 # define TAB_ERROR   "Error -> init philo tab\n"
 
 void	*ft_calloc(size_t nmemb, size_t size);
-void  free_all(t_philo *philo_tab, size_t tab_size);
+void	free_all(t_philo *philo_tab, int tab_size);
 int		print_error(char *str);
 
 #endif
